+++ cups/http-addrlist.c
@@ -1,8 +1,8 @@
 /*
  * HTTP address list routines for CUPS.
  *
- * Copyright 2007-2018 by Apple Inc.
- * Copyright 1997-2007 by Easy Software Products, all rights reserved.
+ * Copyright © 2007-2021 by Apple Inc.
+ * Copyright © 1997-2007 by Easy Software Products, all rights reserved.
  *
  * Licensed under Apache License v2.0.  See the file "LICENSE" for more
  * information.
@@ -99,6 +99,8 @@ httpAddrConnect2(
 
   if (cancel && *cancel)
     return (NULL);
+
+  httpInitialize();
 
   if (msec <= 0)
     msec = INT_MAX;
@@ -337,6 +339,18 @@ httpAddrConnect2(
 	else if (FD_ISSET(fds[i], &error_set))
 #  endif /* HAVE_POLL */
         {
+#  ifdef __sun
+          // Solaris incorrectly returns errors when you poll() a socket that is
+          // still connecting.  This check prevents us from removing the socket
+          // from the pool if the "error" is EINPROGRESS...
+          int		sockerr;	// Current error on socket
+          socklen_t	socklen = sizeof(sockerr);
+					// Size of error variable
+
+          if (!getsockopt(fds[i], SOL_SOCKET, SO_ERROR, &sockerr, &socklen) && (!sockerr || sockerr == EINPROGRESS))
+            continue;			// Not an error
+#  endif // __sun
+
          /*
           * Error on socket, remove from the "pool"...
           */
@@ -503,6 +517,8 @@ httpAddrGetList(const char *hostname,	/* I - Hostname, IP address, or NULL for p
 #  endif /* AF_INET6 */
 	                 family == AF_INET ? "INET" : "???", service);
 #endif /* DEBUG */
+
+  httpInitialize();
 
 #ifdef HAVE_RES_INIT
  /*