+++ cups/http.c
@@ -2,7 +2,7 @@
  * HTTP routines for CUPS.
  *
  * Copyright © 2021 by OpenPrinting.
- * Copyright © 2007-2019 by Apple Inc.
+ * Copyright © 2007-2021 by Apple Inc.
  * Copyright © 1997-2007 by Easy Software Products, all rights reserved.
  *
  * This file contains Kerberos support code, copyright 2006 by
@@ -2422,6 +2422,7 @@ httpReconnect2(http_t *http,		/* I - HTTP connection */
     if (_httpTLSStart(http) != 0)
     {
       httpAddrClose(NULL, http->fd);
+      http->fd = -1;
 
       return (-1);
     }
@@ -2787,6 +2788,7 @@ _httpUpdate(http_t        *http,	/* I - HTTP connection */
       if (_httpTLSStart(http) != 0)
       {
         httpAddrClose(NULL, http->fd);
+        http->fd = -1;
 
 	*status = http->status = HTTP_STATUS_ERROR;
 	return (0);
@@ -4400,14 +4402,19 @@ http_send(http_t       *http,		/* I - HTTP connection */
   }
 
   for (i = 0; i < HTTP_FIELD_MAX; i ++)
+  {
     if ((value = httpGetField(http, i)) != NULL && *value)
     {
       DEBUG_printf(("5http_send: %s: %s", http_fields[i], value));
 
       if (i == HTTP_FIELD_HOST)
       {
-	if (httpPrintf(http, "Host: %s:%d\r\n", value,
-	               httpAddrPort(http->hostaddr)) < 1)
+        // Issue #185: Use "localhost" for the loopback addresses to work
+        // around an Avahi bug...
+        if (httpAddrLocalhost(http->hostaddr))
+          value = "localhost";
+
+	if (httpPrintf(http, "Host: %s:%d\r\n", value, httpAddrPort(http->hostaddr)) < 1)
 	{
 	  http->status = HTTP_STATUS_ERROR;
 	  return (-1);
@@ -4419,6 +4426,7 @@ http_send(http_t       *http,		/* I - HTTP connection */
 	return (-1);
       }
     }
+  }
 
   if (http->cookie)
     if (httpPrintf(http, "Cookie: $Version=0; %s\r\n", http->cookie) < 1)